rename duckdb tool packages
